Autogenerate *_password.h when building if they're not found
This makes it way easier to get started.  Before you had to know how to manually
create the _password.h files for OTA and wifi password after each new checkout.
This change adds logic into build.sh that checks for them and prompts the
user to type in the passwords at build time if they're not found and generates
the files automatically.

After this change when you check out a new copy of the code you'll have to
type in the two passwords on your first build, then shouldn't have to mess with
it anymore after that.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>