Port of some basic animations & animation framework
This is basically just a stripped-down version of the animation framework that I've
written and used in the past for LED controllers.  I've only included a couple of
animations for the sake of simplicity, and because I'm likely going to have to change
the basic API for animations once I am trying to syncronyze them.  This allows a quite
thorough test of the development board and proves that it seems to be working.

When you build this firmware you'll see three different randomized animations playing
in turn on the LEDs.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>