Add brightness selection support
This piggy-backs on the flash-mode button connected to GPIO0 and uses it
as a brightness control.  You can now press the button to toggle through
three different brightness levels.

Note: There was an additional button added to the EVT1 board of GPIO16 that
was intended for this purpose, but it turns out that GPIO16 is the only one
without interrupt support :(  As a happy surprise though, it turns out that
there's no reason I can't use the flash-mode button for both setting the
microcontroller to accept a new program *and* to use it as a brightness
control, so that's what I did here.  I can likely do away with the extra
button (on GPIO16) entirely on the next rev.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>