Transition to PainlessMesh (WIP)
This is the first step to having a good transition to using PainlessMesh
and updating the Animation suite to run using a scheduler.  Right now,
it's not running/compiling right but I need to transition over to another
machine and this is the easiest way to copy it right now.  Oops :]

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>