Reenable PainlessMesh and fix some debug outputs
It's working on the new pro-made PCBs!  Looks like there was some kind
of HW fault on my handmade dev board that was causing resets when I
enabled PainlessMesh.  Now that I have higher quality PCBs it works, so
this commit basically just enables the commented out networking code
and adds in/fixes a few debugging prints to make it easy to prove/test
that the mesh network is working.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>