Actual first time working, lol
I was mistaken before -- this is the first time it's working.  This fixed a problem where
things weren't working if the program never left the loop() function. I didn't think that
would be a prolem, but apparently it was.  Now that we get into the loop() function
things are working!  I was able to push my first OTA update after triggering it o reboot
into OTA mode. Yay!  Now I just gotta clean it up and make it look good, haha

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>