Get broadcast programming working in earnest
This is the one!  I just updated four units all at once over OTA with this!
This change makes it so each unit broadcasts the PROG message for a while before
rebooting regardless of if it was triggered over mesh networking or over
the serial connection.  This seems to work much better and propogates the
PROG message to all four of my testing units.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>